requirements.txt
torch>=1.8.1
pandas=1.4
dgl
torch-geometric
tqdm
networkx
geopandas
faiss-cpu
scikit-learn
seaborn
geatpy=2.7.0
numpy=1.25
geopy
dgl-cu111=0.6.1
gensim
pynvml
psutil
tensorboard
numpy_indexed